backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
pydantic>=2.0.0
sqlalchemy
langchain
reportlab
openpyxl
pydantic-settings
alembic
cryptography
google-genai